+++ CHANGELOG.md
@@ -13,6 +13,7 @@ y este proyecto sigue [Versionado Semántico](https://semver.org/lang/es/).
 - Entidad [Category](src/main/kotlin/com/jeremiascortes/LoQueTengo/backend/entity/Category.kt)
 - Entidad [Product](src/main/kotlin/com/jeremiascortes/LoQueTengo/backend/entity/Product.kt)
 - Entidad [Purchase](src/main/kotlin/com/jeremiascortes/LoQueTengo/backend/entity/Purchase.kt)
+- Entidad [PurchaseDetail](src/main/kotlin/com/jeremiascortes/LoQueTengo/backend/entity/PurchaseDetail.kt)
 
 ## [0.0.1] - 2026-07-03